Adds README and fixes one bug

+++ vim/install.sh
@@ -1,9 +1,36 @@
+#!/bin/bash
+
 #####################################
 # Install basic vim settings
 #####################################
 
-cp -v  `pwd`/vimrc ~/.vimrc
+if [ -f ~/.vimrc ]; then
+	echo "This will install a vim setup."
+	echo "You already have a ~/.vimrc, and it will be"
+	echo "overwritten."
+	echo
+	read -p "Press ENTER to continue, CTRL+C to abort."
+fi
+
+# Create vim directory if it does not exist
+if [ ! -d ~/.vim ]; then
+	mkdir -p ~/.vim/bundle
+fi
+
+# Use interactive copying
+echo
+echo Setting up ~/.vimrc
+echo
+cp -v `pwd`/vimrc ~/.vimrc
+
+echo
+echo Setting up ~/.vim
+echo
 cp -rvuT `pwd`/vim ~/.vim
+
+echo
+echo Setting up ~/.fonts
+echo
 cp -rv `pwd`/fonts ~/.fonts
 
 #####################################
@@ -27,7 +54,7 @@ https://github.com/ervandew/supertab.git
 https://github.com/tomtom/tlib_vim.git
 https://github.com/MarcWeber/vim-addon-mw-utils.git
 https://github.com/garbas/vim-snipmate.git
-https://github.com/honza/snipmate-snippets.git
+https://github.com/honza/vim-snippets
 https://github.com/skwp/vim-powerline.git
 https://github.com/Rip-Rip/clang_complete.git
 https://github.com/scrooloose/nerdtree.git
@@ -35,7 +62,11 @@ https://github.com/xuhdev/SingleCompile.git
 https://github.com/vim-scripts/taglist.vim.git
 https://github.com/jistr/vim-nerdtree-tabs.git
 https://github.com/skwp/vim-colors-solarized.git
-http://github.com/sjl/gundo.vim.git"
+https://github.com/sjl/gundo.vim.git"
+
+echo
+echo Installing bundles
+echo
 
 if [ -d ~/.vim/bundle ]; then
 	cd ~/.vim/bundle
